Rename config/api.js to config/url.js
